# controller_openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Problemologist Controller",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/script-tools/validate": {
      "post": {
        "tags": [
          "script-tools"
        ],
        "summary": "Validate Script",
        "operationId": "validate_script_api_script_tools_validate_post",
        "parameters": [
          {
            "name": "x-session-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Session-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScriptToolRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BenchmarkToolResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/script-tools/simulate": {
      "post": {
        "tags": [
          "script-tools"
        ],
        "summary": "Simulate Script",
        "operationId": "simulate_script_api_script_tools_simulate_post",
        "parameters": [
          {
            "name": "x-session-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Session-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScriptToolRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BenchmarkToolResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/script-tools/verify": {
      "post": {
        "tags": [
          "script-tools"
        ],
        "summary": "Verify Script",
        "operationId": "verify_script_api_script_tools_verify_post",
        "parameters": [
          {
            "name": "x-session-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Session-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScriptToolRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BenchmarkToolResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/script-tools/preview": {
      "post": {
        "tags": [
          "script-tools"
        ],
        "summary": "Preview Script",
        "operationId": "preview_script_api_script_tools_preview_post",
        "parameters": [
          {
            "name": "x-session-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Session-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScriptToolRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreviewDesignResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/script-tools/submit": {
      "post": {
        "tags": [
          "script-tools"
        ],
        "summary": "Submit Script",
        "operationId": "submit_script_api_script_tools_submit_post",
        "parameters": [
          {
            "name": "x-session-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Session-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScriptToolRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BenchmarkToolResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/script-tools/validate": {
      "post": {
        "tags": [
          "script-tools"
        ],
        "summary": "Validate Script",
        "operationId": "validate_script_script_tools_validate_post",
        "parameters": [
          {
            "name": "x-session-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Session-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScriptToolRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BenchmarkToolResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/script-tools/simulate": {
      "post": {
        "tags": [
          "script-tools"
        ],
        "summary": "Simulate Script",
        "operationId": "simulate_script_script_tools_simulate_post",
        "parameters": [
          {
            "name": "x-session-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Session-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScriptToolRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BenchmarkToolResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/script-tools/verify": {
      "post": {
        "tags": [
          "script-tools"
        ],
        "summary": "Verify Script",
        "operationId": "verify_script_script_tools_verify_post",
        "parameters": [
          {
            "name": "x-session-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Session-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScriptToolRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BenchmarkToolResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/script-tools/preview": {
      "post": {
        "tags": [
          "script-tools"
        ],
        "summary": "Preview Script",
        "operationId": "preview_script_script_tools_preview_post",
        "parameters": [
          {
            "name": "x-session-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Session-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScriptToolRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PreviewDesignResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/script-tools/submit": {
      "post": {
        "tags": [
          "script-tools"
        ],
        "summary": "Submit Script",
        "operationId": "submit_script_script_tools_submit_post",
        "parameters": [
          {
            "name": "x-session-id",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "title": "X-Session-Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScriptToolRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BenchmarkToolResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AgentName": {
        "type": "string",
        "enum": [
          "benchmark_planner",
          "benchmark_plan_reviewer",
          "benchmark_coder",
          "benchmark_reviewer",
          "engineer_planner",
          "engineer_coder",
          "engineer_plan_reviewer",
          "electronics_planner",
          "electronics_engineer",
          "electronics_reviewer",
          "engineer_execution_reviewer",
          "skill_agent",
          "git_agent",
          "cots_search",
          "journalling_agent",
          "steer"
        ],
        "title": "AgentName",
        "description": "Canonical agent identifiers used across API, evals, and orchestration."
      },
...
      "BenchmarkToolResponse": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "confidence": {
            "type": "string",
            "enum": [
              "low",
              "medium",
              "high",
              "approximate"
            ],
            "title": "Confidence",
            "default": "high"
          },
          "artifacts": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/SimulationArtifacts"
              },
              {
                "type": "null"
              }
            ]
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/BaseEvent"
            },
            "type": "array",
            "title": "Events"
          }
        },
        "type": "object",
        "required": [
          "success",
          "message"
        ],
        "title": "BenchmarkToolResponse",
        "description": "Response from a benchmark tool."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PreviewRenderingType": {
        "type": "string",
        "enum": [
          "rgb",
          "depth",
          "segmentation"
        ],
        "title": "PreviewRenderingType",
        "description": "Supported single-view preview modalities."
      },
...
      "ScriptToolRequest": {
        "properties": {
          "script_path": {
            "type": "string",
            "title": "Script Path",
            "default": "script.py"
          },
          "agent_role": {
            "$ref": "#/components/schemas/AgentName",
            "default": "engineer_coder"
          },
          "backend": {
            "$ref": "#/components/schemas/SimulatorBackendType"
          },
          "smoke_test_mode": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Smoke Test Mode"
          },
          "bundle_base64": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bundle Base64"
          },
          "pitch": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              }
            ],
            "title": "Pitch",
            "default": -45.0
          },
          "yaw": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "items": {
                  "type": "number"
                },
                "type": "array"
              }
            ],
            "title": "Yaw",
            "default": 45.0
          },
          "rgb": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rgb"
          },
          "depth": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Depth"
          },
          "segmentation": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Segmentation"
          },
          "drafting": {
            "type": "boolean",
            "title": "Drafting",
            "default": false
          },
          "rendering_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PreviewRenderingType"
              },
              {
                "type": "null"
              }
            ]
          },
          "reviewer_stage": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "benchmark_reviewer",
                  "engineering_execution_reviewer",
                  "electronics_reviewer"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewer Stage"
          },
          "jitter_range": {
            "anyOf": [
              {
                "prefixItems": [
                  {
                    "type": "number"
                  },
                  {
                    "type": "number"
                  },
                  {
                    "type": "number"
                  }
                ],
                "type": "array",
                "maxItems": 3,
                "minItems": 3
              },
              {
                "type": "null"
              }
            ],
            "title": "Jitter Range"
          },
          "num_scenes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Num Scenes"
          },
          "duration": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration"
          },
          "seed": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Seed"
          },
          "episode_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Episode Id"
          }
        },
        "type": "object",
        "title": "ScriptToolRequest"
      },
...
      "SimulatorBackendType": {
        "type": "string",
        "enum": [
          "MUJOCO",
          "GENESIS"
        ],
        "title": "SimulatorBackendType"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}